requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
aiosqlite>=0.20
argon2-cffi>=23.1
python-jose[cryptography]>=3.3
python-dotenv>=1.0
httpx>=0.27
mcp>=1.0
